requirements.txt
Pillow==7.0.0
pkg-resources==0.0.0
